% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form_sequense.R
\name{s_Sequence_low}
\alias{s_Sequence_low}
\title{"1" and "0" sequense receiving}
\usage{
s_Sequence_low(sample, operand)
}
\arguments{
\item{sample}{- numeric vector}

\item{operand}{- constant numeric value}
}
\value{
numeric the estimated amount of good honey
}
\description{
Comparing inputting numeric sequense with constant operand.
Receiving logic sequense of comparation
Then transforming into "1" and "0" sequense
}
\details{
Functions return "1" and "0" string, type char
}
\examples{
const <- 5
seq <- c(1:10)
s_Sequence_up(seq, const)
#"0000011111"
}
