% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control_tests.R
\name{Shewchart_5}
\alias{Shewchart_5}
\title{Test №5}
\usage{
Shewchart_5(table, sample, average, sigma)
}
\description{
12 out of 3 consecutive points are more than 2 sigmas
from the center line in the same direction.
}
