% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control_tests.R
\name{Shewchart_6}
\alias{Shewchart_6}
\title{Test №6}
\usage{
Shewchart_6(table, sample, average, sigma)
}
\description{
4 out of 5 consecutive points are more than 1 sigma
from the center line in the same direction.
}
